+++ src/main/java/domain/piece/Sang.java
@@ -2,13 +2,12 @@ package domain.piece;
 
 import domain.Move;
 import domain.Moves;
-import domain.Position;
 import domain.Team;
 import java.util.List;
 
-public class Sang extends Piece {
+public class Sang extends FixedMovePiece {
 
-    private static final List<Moves> moves = List.of(
+    private static final List<Moves> movesOptions = List.of(
             Moves.create(Move.FRONT, Move.FRONT_LEFT, Move.FRONT_LEFT),
             Moves.create(Move.FRONT, Move.FRONT_RIGHT, Move.FRONT_RIGHT),
             Moves.create(Move.BACK, Move.BACK_LEFT, Move.BACK_LEFT),
@@ -24,12 +23,7 @@ public class Sang extends Piece {
     }
 
     @Override
-    public List<Position> calculatePath(Position startPosition, Position targetPosition) {
-        Moves possibleMoves = moves.stream()
-                .filter(moves -> moves.isPossibleToArrive(startPosition, targetPosition))
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("이 위치로 이동할 수 없습니다."));
-
-        return possibleMoves.convertToPath(startPosition);
+    protected List<Moves> getMovesOptions() {
+        return movesOptions;
     }
 }
